requirements2.txt
djangorestframework==3.14.0
filelock==3.9.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
ndg-httpsclient==0.5.1
phonenumberslite==8.13.5
rfc3986==1.5.0
six==1.16.0
small-small-hr==0.4.2
sniffio==1.3.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
tzlocal==4.2